use int instead int64 for status code

+++ domain/api_domain.go
@@ -1,7 +1,7 @@
 package domain
 
 type APIResponse struct {
-	StatusCode int64       `json:"status_code"`
+	StatusCode int         `json:"status_code"`
 	Message    string      `json:"message"`
 	Data       interface{} `json:"data,omitempty"`
 }
